// packages/db-client/generated/kurrentdb/protocols/v2/registry/validation_pb.d.ts
export class SchemaCompatibilityError extends jspb.Message { 
    getKind(): SchemaCompatibilityErrorKind;
    setKind(value: SchemaCompatibilityErrorKind): SchemaCompatibilityError;
    getDetails(): string;
    setDetails(value: string): SchemaCompatibilityError;

    hasPropertyPath(): boolean;
    clearPropertyPath(): void;
    getPropertyPath(): string | undefined;
    setPropertyPath(value: string): SchemaCompatibilityError;

    hasOriginalType(): boolean;
    clearOriginalType(): void;
    getOriginalType(): string | undefined;
    setOriginalType(value: string): SchemaCompatibilityError;

    hasNewType(): boolean;
    clearNewType(): void;
    getNewType(): string | undefined;
    setNewType(value: string): SchemaCompatibilityError;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): SchemaCompatibilityError.AsObject;
    static toObject(includeInstance: boolean, msg: SchemaCompatibilityError): SchemaCompatibilityError.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: SchemaCompatibilityError, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): SchemaCompatibilityError;
    static deserializeBinaryFromReader(message: SchemaCompatibilityError, reader: jspb.BinaryReader): SchemaCompatibilityError;
}

export namespace SchemaCompatibilityError {
    export type AsObject = {
        kind: SchemaCompatibilityErrorKind,
        details: string,
        propertyPath?: string,
        originalType?: string,
        newType?: string,
    }
}
// ...
export enum SchemaCompatibilityErrorKind {
    SCHEMA_COMPATIBILITY_ERROR_KIND_UNSPECIFIED = 0,
    SCHEMA_COMPATIBILITY_ERROR_KIND_MISSING_REQUIRED_PROPERTY = 1,
    SCHEMA_COMPATIBILITY_ERROR_KIND_INCOMPATIBLE_TYPE_CHANGE = 2,
    SCHEMA_COMPATIBILITY_ERROR_KIND_OPTIONAL_TO_REQUIRED = 3,
    SCHEMA_COMPATIBILITY_ERROR_KIND_NEW_REQUIRED_PROPERTY = 4,
    SCHEMA_COMPATIBILITY_ERROR_KIND_REMOVED_PROPERTY = 5,
    SCHEMA_COMPATIBILITY_ERROR_KIND_ARRAY_TYPE_INCOMPATIBILITY = 6,
    SCHEMA_COMPATIBILITY_ERROR_KIND_DATA_FORMAT_MISMATCH = 7,
}